[transitions] move transitions from config package to transitions package
