+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/tasks/CheckForNewTaskActionTransition.java
@@ -1,13 +1,15 @@
-package com.community.tools.util.statemachie.actions.transitions.configs.questions;
+package com.community.tools.util.statemachie.actions.transitions.tasks;
 
-import static com.community.tools.util.statemachie.Event.QUESTION_SECOND;
-import static com.community.tools.util.statemachie.State.FIRST_QUESTION;
-import static com.community.tools.util.statemachie.State.SECOND_QUESTION;
+import static com.community.tools.util.statemachie.Event.GET_THE_NEW_TASK;
+import static com.community.tools.util.statemachie.State.CHECK_FOR_NEW_TASK;
+import static com.community.tools.util.statemachie.State.GOT_THE_FIRST_TASK;
 
 import com.community.tools.service.slack.SlackService;
 import com.community.tools.util.statemachie.Event;
 import com.community.tools.util.statemachie.State;
 import com.community.tools.util.statemachie.actions.Transition;
+import java.util.Arrays;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.statemachine.StateContext;
@@ -16,13 +18,13 @@ import org.springframework.statemachine.annotation.WithStateMachine;
 import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
 
 @WithStateMachine
-public class SecondQuestionActionTransition implements Transition {
-
-  @Value("${secondQuestion}")
-  private String secondQuestion;
+public class CheckForNewTaskActionTransition implements Transition {
 
   @Autowired
   private SlackService slackService;
+
+  @Value("${tasksForUsers}")
+  private String[] tasksForUsers;
 
   @Autowired
   private Action<State, Event> errorAction;
@@ -32,15 +34,21 @@ public class SecondQuestionActionTransition implements Transition {
       StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
     transitions
         .withExternal()
-        .source(FIRST_QUESTION)
-        .target(SECOND_QUESTION)
-        .event(QUESTION_SECOND)
+        .source(GOT_THE_FIRST_TASK)
+        .target(CHECK_FOR_NEW_TASK)
+        .event(GET_THE_NEW_TASK)
         .action(this, errorAction);
   }
 
   @Override
   public void execute(StateContext<State, Event> stateContext) {
+    List<String> tasksList = Arrays.asList(tasksForUsers);
+
+    int i = (Integer) stateContext.getExtendedState().getVariables().get("taskNumber");
+    String taskMessage =
+        "[{\"type\": \"section\",\"text\": {\"type\": \"mrkdwn\",\"text\": \"Here is your next <https://github.com/Broscorp-net/traineeship/tree/master/module1/src/main/java/net/broscorp/"
+            + tasksList.get(i) + "|TASK>.\"}}]";
     String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendBlocksMessage(slackService.getUserById(user), secondQuestion);
+    slackService.sendBlocksMessage(slackService.getUserById(user), taskMessage);
   }
 }
