+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/tasks/GetTheFirstTaskActionTransition.java
@@ -1,8 +1,8 @@
-package com.community.tools.util.statemachie.actions.transitions.configs.questions;
+package com.community.tools.util.statemachie.actions.transitions.tasks;
 
-import static com.community.tools.util.statemachie.Event.QUESTION_FIRST;
-import static com.community.tools.util.statemachie.State.FIRST_QUESTION;
-import static com.community.tools.util.statemachie.State.NEW_USER;
+import static com.community.tools.util.statemachie.Event.GET_THE_FIRST_TASK;
+import static com.community.tools.util.statemachie.State.ADDED_GIT;
+import static com.community.tools.util.statemachie.State.GOT_THE_FIRST_TASK;
 
 import com.community.tools.service.slack.SlackService;
 import com.community.tools.util.statemachie.Event;
@@ -16,10 +16,10 @@ import org.springframework.statemachine.annotation.WithStateMachine;
 import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
 
 @WithStateMachine
-public class FirstQuestionActionTransition implements Transition {
+public class GetTheFirstTaskActionTransition implements Transition {
 
-  @Value("${firstQuestion}")
-  private String firstQuestion;
+  @Value("${getFirstTask}")
+  private String getFirstTask;
 
   @Autowired
   private SlackService slackService;
@@ -28,19 +28,19 @@ public class FirstQuestionActionTransition implements Transition {
   private Action<State, Event> errorAction;
 
   @Override
-  public void execute(StateContext<State, Event> stateContext) {
-    String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendBlocksMessage(slackService.getUserById(user), firstQuestion);
-  }
-
-  @Override
   public void configure(
       StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
     transitions
         .withExternal()
-        .source(NEW_USER)
-        .target(FIRST_QUESTION)
-        .event(QUESTION_FIRST)
+        .source(ADDED_GIT)
+        .target(GOT_THE_FIRST_TASK)
+        .event(GET_THE_FIRST_TASK)
         .action(this, errorAction);
+  }
+
+  @Override
+  public void execute(StateContext<State, Event> stateContext) {
+    String user = stateContext.getExtendedState().getVariables().get("id").toString();
+    slackService.sendBlocksMessage(slackService.getUserById(user), getFirstTask);
   }
 }
