+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/verifications/AgreeLicenseActionTransition.java
@@ -1,8 +1,8 @@
-package com.community.tools.util.statemachie.actions.transitions.configs.verifications;
+package com.community.tools.util.statemachie.actions.transitions.verifications;
 
-import static com.community.tools.util.statemachie.Event.DID_NOT_PASS_VERIFICATION_GIT_LOGIN;
+import static com.community.tools.util.statemachie.Event.AGREE_LICENSE;
 import static com.community.tools.util.statemachie.State.AGREED_LICENSE;
-import static com.community.tools.util.statemachie.State.CHECK_LOGIN;
+import static com.community.tools.util.statemachie.State.INFORMATION_CHANNELS;
 
 import com.community.tools.service.slack.SlackService;
 import com.community.tools.util.statemachie.Event;
@@ -16,30 +16,31 @@ import org.springframework.statemachine.annotation.WithStateMachine;
 import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
 
 @WithStateMachine
-public class DidNotPassVerificationGitLoginTrans implements Transition {
+public class AgreeLicenseActionTransition implements Transition {
+
+  @Value("${addGitName}")
+  private String addGitName;
+
+  @Autowired
+  private SlackService slackService;
 
   @Autowired
   private Action<State, Event> errorAction;
-  @Value("${answeredNoDuringVerification}")
-  private String answeredNoDuringVerification;
-  @Autowired
-  private SlackService slackService;
 
   @Override
   public void configure(
       StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
     transitions
         .withExternal()
-        .source(CHECK_LOGIN)
+        .source(INFORMATION_CHANNELS)
         .target(AGREED_LICENSE)
-        .event(DID_NOT_PASS_VERIFICATION_GIT_LOGIN)
+        .event(AGREE_LICENSE)
         .action(this, errorAction);
   }
 
   @Override
   public void execute(StateContext<State, Event> stateContext) {
     String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendPrivateMessage(slackService.getUserById(user),
-        answeredNoDuringVerification);
+    slackService.sendBlocksMessage(slackService.getUserById(user), addGitName);
   }
 }
