+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/verifications/DidNotPassVerificationGitLoginTrans.java
@@ -1,8 +1,8 @@
-package com.community.tools.util.statemachie.actions.transitions.configs.verifications;
+package com.community.tools.util.statemachie.actions.transitions.verifications;
 
-import static com.community.tools.util.statemachie.Event.AGREE_LICENSE;
+import static com.community.tools.util.statemachie.Event.DID_NOT_PASS_VERIFICATION_GIT_LOGIN;
 import static com.community.tools.util.statemachie.State.AGREED_LICENSE;
-import static com.community.tools.util.statemachie.State.INFORMATION_CHANNELS;
+import static com.community.tools.util.statemachie.State.CHECK_LOGIN;
 
 import com.community.tools.service.slack.SlackService;
 import com.community.tools.util.statemachie.Event;
@@ -16,31 +16,30 @@ import org.springframework.statemachine.annotation.WithStateMachine;
 import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
 
 @WithStateMachine
-public class AgreeLicenseActionTransition implements Transition {
-
-  @Value("${addGitName}")
-  private String addGitName;
-
-  @Autowired
-  private SlackService slackService;
+public class DidNotPassVerificationGitLoginTrans implements Transition {
 
   @Autowired
   private Action<State, Event> errorAction;
+  @Value("${answeredNoDuringVerification}")
+  private String answeredNoDuringVerification;
+  @Autowired
+  private SlackService slackService;
 
   @Override
   public void configure(
       StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
     transitions
         .withExternal()
-        .source(INFORMATION_CHANNELS)
+        .source(CHECK_LOGIN)
         .target(AGREED_LICENSE)
-        .event(AGREE_LICENSE)
+        .event(DID_NOT_PASS_VERIFICATION_GIT_LOGIN)
         .action(this, errorAction);
   }
 
   @Override
   public void execute(StateContext<State, Event> stateContext) {
     String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendBlocksMessage(slackService.getUserById(user), addGitName);
+    slackService.sendPrivateMessage(slackService.getUserById(user),
+        answeredNoDuringVerification);
   }
 }
